rename AbstractCreateKernel and AbstractCreateSymmetricKernel to AbstractCreateGaussianKernel and AbstractCreateSymmetricGaussianKernel
