+++ src/main/java/net/imagej/ops/create/kernelGauss/CreateKernelGaussSymmetric.java
@@ -31,7 +31,7 @@
 package net.imagej.ops.create.kernelGauss;
 
 import net.imagej.ops.Ops;
-import net.imagej.ops.create.AbstractCreateSymmetricKernel;
+import net.imagej.ops.create.AbstractCreateSymmetricGaussianKernel;
 import net.imglib2.img.Img;
 import net.imglib2.type.numeric.ComplexType;
 
@@ -46,7 +46,7 @@ import org.scijava.plugin.Plugin;
  */
 @Plugin(type = Ops.Create.KernelGauss.class, priority = Priority.HIGH_PRIORITY)
 public class CreateKernelGaussSymmetric<T extends ComplexType<T>> extends
-	AbstractCreateSymmetricKernel<T> implements Ops.Create.KernelGauss
+	AbstractCreateSymmetricGaussianKernel<T> implements Ops.Create.KernelGauss
 {
 
 	@Override
